cleanup of main code and structures.
Separation of concerns for more readability. Separation of files in appropriate folders. Implementation of structure handlerService to contain common values/structures used amongst the handlerService functions. I would have also liked to implement an interface but there is no need for it as of yet.
