+++ cmd/api/models/models.go
@@ -1,0 +1,7 @@
+package models
+
+type Message struct {
+	Sender   string `json:"sender"`
+	Receiver string `json:"receiver"`
+	Message  string `json:"message"`
+}
